// src/BackEnd_Circulo_Lectura/src/interfaces/modelosApp/modelosApp.ts
export interface LibroResumen {
	id_libro: number;
	titulo_libro: string;
	autores: Array<{ nombre_autor: string; apellido_autor: string }>;
	calificacionPromedio?: number;
}

export interface ListaApp {
	id_lista: number;
	id_usuarioCreador: number;
	nombre_lista: string;
	nombreCreador: string;
	categorias: string[];
	librosPortada: number[]; // Solo los 2-3 ids para portada
	totalLibros: number;
	totalSeguidores: number;
	totalMeGusta: number;
	puntuacionPromedio?: number | null;
	totalCalificaciones?: number;
	distribucion?: { nota: number; cantidad: number; frecuencia: number }[];
	descripcion_lista?: string;
}